+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Enhancements to PayPal customer recommendations
   - Create a session and get recommendations in one call
   - Hash customer email and phone number
+- Add `business`, `consumer`, `corporate`, and `purchase` from the bin data to credit card responses
 
 ## 4.27.0